chore: rename upsert SQL command module to write SQL
