+++ cli/zeus.js
@@ -30,7 +30,7 @@ const { run: runQA } = require('../src/cli/commands/qaCommand');
 const { runAssessRisk } = require('../src/cli/commands/assessRiskCommand');
 const { runGenerateTest } = require('../src/cli/commands/generateTestCommand');
 const { runGenerateChecklist } = require('../src/cli/commands/generateChecklistCommand');
-const { runUpsertSql, runInsertSql, runUpdateSql } = require('../src/cli/commands/upsertSqlCommand');
+const { runUpsertSql, runInsertSql, runUpdateSql } = require('../src/cli/commands/writeSqlCommand');
 const { runInspectObject } = require('../src/cli/commands/inspectObjectCommand');
 const { run: runTestRun } = require('../src/cli/commands/testRunCommand');
 const { runBridge } = require('../src/cli/commands/bridgeCommand');
