+++ src/cli/commands/writeSqlCommand.js
